Initial Checkin
This is an initial checkin for weatherCheck
The idea of weatherCheck is to retrieve weather data from Weather
Underground.com's API
WU is a great service and allows for a large amount of historical data
as well as current and forecasted weather. (Up to 3 days ahead, and
historical pulls are 1 day/call, know what you're asking for!)

NOTE: ---------- Weather Underground is a pay-to-play service after 500
API calls/day with a limit of 10 calls/minute
It is free to signup and get your own API key and this script accepts
keys as CLI args. Please sign up for your own key!

This checkin is fully usable from the CLI.
Future additions will include extra features like the amount of days to
look back, taking 'city, state' instead of just zipcode,  detailed
logging, and a few others.